chore: Moved AddArgs to inter.py
